refactor(message-core): resource exports have to be classes and not instances
