+++ packages/message-core/src/index.js
@@ -1,3 +1,6 @@
 import AvMessage from './AvMessage';
+import * as Constants from './Constants';
+
+export { Constants };
 
 export default new AvMessage();
